third_party/blink/web_tests/fast/dom/simultaneouslyRegsiteredTimerFireOrder-expected.txt

This page verifies that timers with equivalent fire times fire in the order they were registered. If the test passes, you'll see a series of PASS messages below.

PASS: firedTimers[0] should be 0 and is.
PASS: firedTimers[1] should be 1 and is.
PASS: firedTimers[2] should be 2 and is.
PASS: firedTimers[3] should be 3 and is.
PASS: firedTimers[4] should be 4 and is.
PASS: firedTimers[5] should be 5 and is.
PASS: firedTimers[6] should be 6 and is.
PASS: firedTimers[7] should be 7 and is.
PASS: firedTimers[8] should be 8 and is.
PASS: firedTimers[9] should be 9 and is.
PASS: firedTimers[10] should be 10 and is.
PASS: firedTimers[11] should be 11 and is.
PASS: firedTimers[12] should be 12 and is.
PASS: firedTimers[13] should be 13 and is.
PASS: firedTimers[14] should be 14 and is.
PASS: firedTimers[15] should be 15 and is.
PASS: firedTimers[16] should be 16 and is.
PASS: firedTimers[17] should be 17 and is.
PASS: firedTimers[18] should be 18 and is.
PASS: firedTimers[19] should be 19 and is.
PASS: firedTimers[20] should be 20 and is.
PASS: firedTimers[21] should be 21 and is.
PASS: firedTimers[22] should be 22 and is.
PASS: firedTimers[23] should be 23 and is.
PASS: firedTimers[24] should be 24 and is.
PASS: firedTimers[25] should be 25 and is.
PASS: firedTimers[26] should be 26 and is.
PASS: firedTimers[27] should be 27 and is.
PASS: firedTimers[28] should be 28 and is.
PASS: firedTimers[29] should be 29 and is.
PASS: firedTimers[30] should be 30 and is.
PASS: firedTimers[31] should be 31 and is.
PASS: firedTimers[32] should be 32 and is.
PASS: firedTimers[33] should be 33 and is.
PASS: firedTimers[34] should be 34 and is.
PASS: firedTimers[35] should be 35 and is.
PASS: firedTimers[36] should be 36 and is.
PASS: firedTimers[37] should be 37 and is.
PASS: firedTimers[38] should be 38 and is.
PASS: firedTimers[39] should be 39 and is.
PASS: firedTimers[40] should be 40 and is.
PASS: firedTimers[41] should be 41 and is.
PASS: firedTimers[42] should be 42 and is.
PASS: firedTimers[43] should be 43 and is.
PASS: firedTimers[44] should be 44 and is.
PASS: firedTimers[45] should be 45 and is.
PASS: firedTimers[46] should be 46 and is.
PASS: firedTimers[47] should be 47 and is.
PASS: firedTimers[48] should be 48 and is.
PASS: firedTimers[49] should be 49 and is.
PASS: firedTimers[50] should be 50 and is.
PASS: firedTimers[51] should be 51 and is.
PASS: firedTimers[52] should be 52 and is.
PASS: firedTimers[53] should be 53 and is.
PASS: firedTimers[54] should be 54 and is.
PASS: firedTimers[55] should be 55 and is.
PASS: firedTimers[56] should be 56 and is.
PASS: firedTimers[57] should be 57 and is.
PASS: firedTimers[58] should be 58 and is.
PASS: firedTimers[59] should be 59 and is.
PASS: firedTimers[60] should be 60 and is.
PASS: firedTimers[61] should be 61 and is.
PASS: firedTimers[62] should be 62 and is.
PASS: firedTimers[63] should be 63 and is.
PASS: firedTimers[64] should be 64 and is.
PASS: firedTimers[65] should be 65 and is.
PASS: firedTimers[66] should be 66 and is.
PASS: firedTimers[67] should be 67 and is.
PASS: firedTimers[68] should be 68 and is.
PASS: firedTimers[69] should be 69 and is.
PASS: firedTimers[70] should be 70 and is.
PASS: firedTimers[71] should be 71 and is.
PASS: firedTimers[72] should be 72 and is.
PASS: firedTimers[73] should be 73 and is.
PASS: firedTimers[74] should be 74 and is.
PASS: firedTimers[75] should be 75 and is.
PASS: firedTimers[76] should be 76 and is.
PASS: firedTimers[77] should be 77 and is.
PASS: firedTimers[78] should be 78 and is.
PASS: firedTimers[79] should be 79 and is.
PASS: firedTimers[80] should be 80 and is.
PASS: firedTimers[81] should be 81 and is.
PASS: firedTimers[82] should be 82 and is.
PASS: firedTimers[83] should be 83 and is.
PASS: firedTimers[84] should be 84 and is.
PASS: firedTimers[85] should be 85 and is.
PASS: firedTimers[86] should be 86 and is.
PASS: firedTimers[87] should be 87 and is.
PASS: firedTimers[88] should be 88 and is.
PASS: firedTimers[89] should be 89 and is.
PASS: firedTimers[90] should be 90 and is.
PASS: firedTimers[91] should be 91 and is.
PASS: firedTimers[92] should be 92 and is.
PASS: firedTimers[93] should be 93 and is.
PASS: firedTimers[94] should be 94 and is.
PASS: firedTimers[95] should be 95 and is.
PASS: firedTimers[96] should be 96 and is.
PASS: firedTimers[97] should be 97 and is.
PASS: firedTimers[98] should be 98 and is.
PASS: firedTimers[99] should be 99 and is.
